tests/e2e/requirements.txt
# E2E Test Dependencies for OAuth2 SASL Plugin
requests>=2.28.0
PyJWT>=2.6.0
cryptography>=3.4.8
imaplib2>=3.6
